change variable type of divide.cpp to accomodate for decimal values

+++ 20-08-2026/impcalc/Operations/divide.cpp
@@ -1,8 +1,8 @@
 #include <iostream>
-int divide(int a, int b) {
+double divide(int a, int b) {
     if (b == 0) {
         std::cout << "Division by zero is not allowed." << std::endl;
         return 0; // Or handle the error in a way that makes sense for your application
     }
-    return a / b;
+    return static_cast<double>(a) / static_cast<double>(b);
 }
